+++ torchtune/training/memory.py
@@ -52,101 +52,6 @@ def cleanup_before_training() -> None:
     get_torch_device_namespace().reset_peak_memory_stats()
 
 
-# fused_optimizer_in_backward.py
-# import torch
-# from torch.optim import Optimizer, AdamW, Adam, SGD
-
-from torchtune import config
-
-
-class FusedOptimizerInBackward:
-    def __init__(
-        self, base_optimizer_cls: str | torch.optim.Optimizer, params, **kwargs
-    ):
-        self.per_param_optimizers = {}
-        self.param_to_opt = {}
-        self._proxy_optimizer = None  # for schedulers
-
-        self._param_groups = []  # Custom param_groups
-
-        for param in params:
-            if not param.requires_grad:
-                continue
-            opt = eval(base_optimizer_cls)([param], **kwargs)
-            self.per_param_optimizers[param] = opt
-            self.param_to_opt[param] = opt
-            self._param_groups.append(opt.param_groups[0])
-            if self._proxy_optimizer is None:
-                self._proxy_optimizer = opt  # first one used for schedulers
-
-            # Hook to call .step() on this param's optimizer
-            if hasattr(param, "register_post_accumulate_grad_hook"):
-                param.register_post_accumulate_grad_hook(
-                    lambda p=param: self._step_and_clear(p)
-                )
-            else:
-                param.register_hook(lambda grad, p=param: self._step_and_clear(p))
-
-    @classmethod
-    def from_config(cls, cfg_optimizer: "DictConfgi", params):
-        pass
-
-    def _step_and_clear(self, param):
-        opt = self.param_to_opt.get(param, None)
-        if opt is None:
-            return
-        opt.step()
-        param.grad = None
-
-    def step(self, closure=None):
-        if closure is not None:
-            with torch.enable_grad():
-                return closure()
-
-    def zero_grad(self, set_to_none=True):
-        for param, opt in self.per_param_optimizers.items():
-            if param.grad is not None:
-                if set_to_none:
-                    param.grad = None
-                else:
-                    param.grad.detach_()
-                    param.grad.zero_()
-
-    def state_dict(self):
-        return {
-            "per_param": {
-                id(p): opt.state_dict() for p, opt in self.per_param_optimizers.items()
-            },
-            "proxy": (
-                self._proxy_optimizer.state_dict() if self._proxy_optimizer else {}
-            ),
-        }
-
-    def load_state_dict(self, state_dict):
-        for p, opt in self.per_param_optimizers.items():
-            key = id(p)
-            if key in state_dict.get("per_param", {}):
-                opt.load_state_dict(state_dict["per_param"][key])
-        if self._proxy_optimizer and "proxy" in state_dict:
-            self._proxy_optimizer.load_state_dict(state_dict["proxy"])
-
-    @property
-    def param_groups(self):
-        return (
-            self._proxy_optimizer.param_groups
-            if self._proxy_optimizer
-            else self._param_groups
-        )
-
-    @property
-    def first_optimizer(self):
-        return next(iter(self.per_param_optimizers.values()))
-
-    @property
-    def defaults(self):
-        return self._proxy_optimizer.defaults if self._proxy_optimizer else {}
-
-
 class OptimizerInBackwardWrapper:
     """
     A bare-bones class meant for checkpoint save and load for optimizers running
